backend/requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.4.0
python-multipart>=0.0.6

# AI & LLM
groq>=0.4.0

# HTTP & Web
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Security & Authentication  
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Input Validation & Sanitization
bleach>=6.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Logging & Monitoring
structlog>=23.1.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0